refactor: split app logic
