+++ src/App.jsx
@@ -11,12 +11,9 @@ import { TRAP_TYPES } from './trapRules'
 import DiscardModal from './components/DiscardModal'
 import RewardModal from './components/RewardModal'
 import DeveloperModal from './components/DeveloperModal'
-import { createShuffledDeck } from './roomDeck'
 import './App.css'
 import { HERO_TYPES } from './heroData'
 import { GOBLIN_TYPES, randomGoblinType } from './goblinData'
-import { randomTreasure, adaptTreasureItem } from './treasureDeck'
-import { formatFightLogs } from './fightUtils'
 import {
   getRangedTargets,
   getMagicTargets,
@@ -24,118 +21,13 @@ import {
   distanceToTarget,
   distanceMagic,
 } from './boardUtils'
-
-const BOARD_SIZE = 7
-const CENTER = Math.floor(BOARD_SIZE / 2)
-
-const DIRS = ['up', 'down', 'left', 'right']
-
-function directionFromDelta(dr, dc) {
-  if (dr === -1) return 'up'
-  if (dr === 1) return 'down'
-  if (dc === -1) return 'left'
-  if (dc === 1) return 'right'
-  return null
-}
-
-function createEmptyBoard() {
-  return Array.from({ length: BOARD_SIZE }, (_, r) =>
-    Array.from({ length: BOARD_SIZE }, (_, c) => ({
-      row: r,
-      col: c,
-      roomId: null,
-      revealed: false,
-      paths: { up: false, down: false, left: false, right: false },
-      goblin: null,
-      effect: null,
-    }))
-  )
-}
-
-function loadState() {
-  const saved = localStorage.getItem('dungeon-state')
-  if (saved) {
-    try {
-      const parsed = JSON.parse(saved)
-      if (parsed.board) {
-        parsed.board.forEach(row =>
-          row.forEach(tile => {
-            if (!tile.paths) tile.paths = { up: false, down: false, left: false, right: false }
-            if (!('goblin' in tile)) tile.goblin = null
-            if (!('trap' in tile)) tile.trap = null
-            if (tile.trap && typeof tile.trap === 'string') {
-              tile.trap = { ...TRAP_TYPES[tile.trap], id: tile.trap }
-            }
-            if (!('trapResolved' in tile)) tile.trapResolved = false
-            if (!('effect' in tile)) tile.effect = null
-          })
-        )
-      }
-      if (parsed.hero) {
-        const type = parsed.hero.type || 'knight'
-        const base = HERO_TYPES[type]
-        const maxHp = parsed.hero.maxHp ?? base.maxHp ?? base.hp
-        const maxAp = parsed.hero.maxAp ?? base.maxAp ?? base.ap
-        const skill = parsed.hero.skill ?? base.skill
-        const skill2 = parsed.hero.skill2 ?? base.skill2
-        const quote = parsed.hero.quote ?? base.quote
-        parsed.hero = {
-          row: parsed.hero.row,
-          col: parsed.hero.col,
-          skill: typeof skill === 'object' ? { ...skill } : skill,
-          skill2: skill2 ? { ...skill2 } : null,
-          movement: parsed.hero.movement ?? base.movement,
-          icon: parsed.hero.icon ?? base.icon,
-          hp: Math.min(parsed.hero.hp ?? base.hp, maxHp),
-          maxHp,
-         ap: Math.min(parsed.hero.ap ?? base.ap, maxAp),
-         maxAp,
-          heroAction: Math.min(
-            parsed.hero.heroAction ?? base.heroAction ?? 1,
-            base.maxHeroAction ?? 1,
-          ),
-          maxHeroAction: parsed.hero.maxHeroAction ?? base.maxHeroAction ?? 1,
-         defence: parsed.hero.defence ?? base.defence,
-          strengthDice: parsed.hero.strengthDice ?? base.strengthDice,
-          agilityDice: parsed.hero.agilityDice ?? base.agilityDice,
-          magicDice: parsed.hero.magicDice ?? base.magicDice,
-          weapons: parsed.hero.weapons ?? base.weapons.map(w => ({ ...w })),
-          name: base.name,
-          image: base.image,
-          quote,
-          type,
-          offset: parsed.hero.offset ?? { x: 0, y: 0 },
-        }
-      }
-      parsed.encounter = null
-      if (!parsed.trap) parsed.trap = null
-      if (!parsed.discard) parsed.discard = null
-      if (!parsed.reward || !parsed.reward.item) parsed.reward = null
-      return parsed
-    } catch {
-      /* ignore corrupted save */
-    }
-  }
-  const board = createEmptyBoard()
-  board[CENTER][CENTER] = {
-    row: CENTER,
-    col: CENTER,
-    roomId: 'Start',
-    revealed: true,
-    paths: { up: true, down: true, left: true, right: true },
-    goblin: null,
-    effect: null,
-  }
-  return {
-    board,
-    hero: null,
-    deck: createShuffledDeck(),
-    encounter: null,
-    trap: null,
-    discard: null,
-    reward: null,
-  }
-}
+import {
+  BOARD_SIZE,
+  CENTER,
+  directionFromDelta,
+  loadState,
+} from './gameSetup'
+import useEncounterHandlers from './hooks/useEncounterHandlers'
 
 function App() {
   const [state, setState] = useState(loadState)
@@ -434,206 +326,15 @@ function App() {
     [state.board],
   )
 
-  const applyDiceRewards = useCallback(
-    rewards => {
-      if (!rewards || (!rewards.ap && !rewards.hp)) return
-      const rewardParts = []
-      if (rewards.ap) rewardParts.push(`${rewards.ap} ap`)
-      if (rewards.hp) rewardParts.push(`${rewards.hp} hp`)
-      addLog(`Unused dice reward: ${rewardParts.join(' and ')}.`)
-      setState(prev => {
-        const hero = prev.hero
-        if (!hero) return prev
-        const newHero = {
-          ...hero,
-          ap: Math.min(hero.ap + rewards.ap, hero.maxAp),
-          hp: Math.min(hero.hp + rewards.hp, hero.maxHp),
-        }
-        return { ...prev, hero: newHero }
-      })
-    },
-    [addLog],
-  )
-
-  const applySkillCost = useCallback(
-    (cost, title) => {
-      if (!cost) return
-      addLog(`Used ${title} (-${cost} AP).`)
-      setState(prev => {
-        const hero = prev.hero
-        if (!hero) return prev
-        return { ...prev, hero: { ...hero, ap: Math.max(0, hero.ap - cost) } }
-      })
-    },
-    [addLog],
-  )
-
-  const handleFight = useCallback(fightResult => {
-    const logs = []
-    setState(prev => {
-      const { encounter, board } = prev
-      if (!encounter || !fightResult) return prev
-      const result = fightResult
-      const newBoard = board.map(row => row.map(tile => ({ ...tile })))
-      const tile = newBoard[encounter.position.row][encounter.position.col]
-      let newEncounter = {
-        ...encounter,
-        goblin: { ...result.goblin, defence: result.defenceAfter },
-      }
-      let newHero = result.hero
-      let discard = prev.discard
-      let reward = prev.reward
-      tile.goblin = { ...result.goblin, defence: result.defenceAfter }
-      const row = encounter.position.row
-      const col = encounter.position.col
-
-      if (result.goblin.hp <= 0) {
-        tile.goblin = null
-        tile.effect = 'death'
-        newEncounter = null
-        const item = adaptTreasureItem(randomTreasure())
-        newHero = { ...newHero, weapons: [...newHero.weapons, item] }
-        reward = { item, hp: 0 }
-        discard = null
-        setTimeout(() => {
-          setState(p => {
-            const copy = p.board.map(r => r.map(t => ({ ...t })))
-            const t = copy[row][col]
-            if (t.effect === 'death') t.effect = null
-            return { ...p, board: copy }
-          })
-        }, 600)
-      } else if (result.counter) {
-        // end encounter after counterattack
-        newEncounter = null
-      } else {
-        // goblin survived but cannot counterattack
-        newEncounter = null
-      }
-      const updated = {
-        ...prev,
-        board: newBoard,
-        hero: newHero,
-        encounter: newEncounter,
-        reward,
-        discard,
-      }
-      return updated
-    })
-    // create log details after state update
-    if (fightResult) {
-      formatFightLogs(fightResult).forEach(l => logs.push(l))
-      // reward already applied when dice were selected
-    }
-    logs.forEach(addLog)
-  }, [addLog])
-
-  const handleFlee = useCallback(success => {
-    let msg = ''
-    setState(prev => {
-      if (!prev.encounter) return prev
-      const { encounter, board, hero } = prev
-      const newBoard = board.map(row => row.map(tile => ({ ...tile })))
-      let newHero = { ...hero, movement: 0 }
-      let newEncounter = encounter
-      if (success) {
-        newHero.row = encounter.prev.row
-        newHero.col = encounter.prev.col
-        newHero.offset = {
-          x: Math.random() * 40 - 20,
-          y: Math.random() * 40 - 20,
-        }
-        newEncounter = null
-        msg = 'Fled successfully.'
-      } else {
-        const damage = Math.max(1, encounter.goblin.attack - hero.defence)
-        newHero.hp = hero.hp - damage
-        msg = `Failed to flee and took ${damage} damage.`
-      }
-      return { ...prev, board: newBoard, hero: newHero, encounter: newEncounter }
-    })
-    addLog(msg)
-  }, [addLog])
-
-  const handleTrapResolve = useCallback(result => {
-    if (!result) return
-    const {
-      evaded,
-      disarm,
-      evasionRolls = [],
-      disarmRolls = [],
-      evasionRewards = { ap: 0, hp: 0 },
-      rewards = { ap: 0, hp: 0 },
-    } = result
-    let messageParts = []
-    setState(prev => {
-      const { trap, board, hero } = prev
-      if (!trap) return prev
-      const newBoard = board.map(row => row.map(tile => ({ ...tile })))
-      const tile = newBoard[trap.position.row][trap.position.col]
-      let newHero = { ...hero }
-      let reward = prev.reward
-      let discard = null
-      if (!evaded) {
-        newHero.hp = hero.hp - tile.trap.damage
-        messageParts.push(`Hit by trap for ${tile.trap.damage} damage.`)
-      } else {
-        newHero.ap = Math.min(hero.ap + evasionRewards.ap, hero.maxAp)
-        newHero.hp = Math.min(newHero.hp + evasionRewards.hp, hero.maxHp)
-        const rParts = []
-        if (evasionRewards.ap) rParts.push(`${evasionRewards.ap} ap`)
-        if (evasionRewards.hp) rParts.push(`${evasionRewards.hp} hp`)
-        messageParts.push(
-          rParts.length
-            ? `Evaded trap and gained ${rParts.join(' and ')}.`
-            : 'Evaded trap.',
-        )
-      }
-      if (disarm !== undefined) {
-        newHero.ap = Math.max(0, Math.min(hero.maxAp, newHero.ap - 1 + (disarm ? rewards.ap : 0)))
-        if (disarm) {
-          tile.trapResolved = true
-          const item = adaptTreasureItem(randomTreasure())
-          newHero.weapons = [...hero.weapons, item]
-          newHero.hp = Math.min(newHero.hp + tile.trap.reward + (rewards.hp || 0), hero.maxHp)
-          reward = { item, hp: tile.trap.reward + (rewards.hp || 0), ap: rewards.ap }
-          const rewardMsg = []
-          if (tile.trap.reward) rewardMsg.push(`${tile.trap.reward} hp`)
-          if (rewards.ap) rewardMsg.push(`${rewards.ap} ap`)
-          if (rewards.hp) rewardMsg.push(`${rewards.hp} bonus hp`)
-          messageParts.push(`Disarmed trap and gained ${rewardMsg.join(' and ')}.`)
-        } else {
-          newHero.hp = newHero.hp - tile.trap.damage
-          messageParts.push(`Failed to disarm and took ${tile.trap.damage} damage.`)
-        }
-      }
-      return { ...prev, board: newBoard, hero: newHero, trap: null, reward, discard }
-    })
-    const rollParts = []
-    if (evasionRolls.length) rollParts.push(`Evade roll ${evasionRolls.join(', ')} (best ${Math.max(...evasionRolls)})`)
-    if (disarmRolls.length) rollParts.push(`Disarm roll ${disarmRolls.join(', ')} (best ${Math.max(...disarmRolls)})`)
-    const rollMsg = rollParts.join('. ')
-    const msg = messageParts.join(' ')
-    addLog(rollMsg ? `${rollMsg}. ${msg}` : msg)
-  }, [addLog])
-
-  const handleRewardConfirm = useCallback(() => {
-    let msg = 'Collected reward'
-    setState(prev => {
-      if (!prev.reward) return prev
-      let discard = null
-      if (prev.hero.weapons.length > 2) {
-        discard = { items: prev.hero.weapons }
-      }
-      return { ...prev, reward: null, discard }
-    })
-    addLog(msg)
-  }, [addLog])
-
-  const handleDiscardConfirm = useCallback(items => {
-    setState(prev => ({ ...prev, hero: { ...prev.hero, weapons: items }, discard: null }))
-    addLog('Chose items to keep')
-  }, [addLog])
+  const {
+    applyDiceRewards,
+    applySkillCost,
+    handleFight,
+    handleFlee,
+    handleTrapResolve,
+    handleRewardConfirm,
+    handleDiscardConfirm,
+  } = useEncounterHandlers(setState, addLog)
 
   useEffect(() => {
     if (!state.hero) return
